Add support for assigning multiple @since tags to the given item.
phpDocumentor already supports and parses all of the @since tags present in a docblock. This PR simply recognizes that there can be more than one for a single item, and also recognizes that they can have descriptions. Adds a 'description' key to the meta array (if the tag description is set) and assigns multiple @since terms as available.