Generate hook names using parsed Node instead of pretty-printed name.
Previously, the hook-name printer was reyling on the parser's pretty
printer to generate names, and then post-processing the name to
extract string literal values and concatenations.

Unfortunately this left some cases unhandled, specifically the case
where concatenations are joining function calls with string literals.
Any number of other unexpected situations might arise, and there can
be defects in the redundant code attempting to parse PHP syntax.

In this patch the pretty-printed version of the expression is used
only as a fallback in unrecognized situations. Primarily, a direct
encoding from the parsed syntax tree to string is used to rely on
the parser's own handling of syntax, and making it clearer how to
add additional support for other syntaxes.